Change return of data in balance, entry and debit

+++ src/main/java/com/financewebapp/api/model/Entry.java
@@ -23,7 +23,7 @@ public class Entry {
     private String entryName;
 
     @Column(name = "EntryValue")
-    private Float entryValue;
+    private Double entryValue;
 
     @CreationTimestamp
     @Column(name = "EntryDate")
